Moved song compnents to own subdir
